+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased](https://github.com/ixocreate/cache/compare/0.1.0...develop)
 
+### Changed
+- Return-Type of CacheInterface->commit() from void to bool
+
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/cache/compare/master...0.1.0)
 
 ### Changed